refactor: new naming for hooks of datatable
