+++ src/components/DataTable/hooks/index.ts
@@ -1,4 +1,4 @@
-export * from './useDataTableCustomCells';
-export * from './useTablePagination';
-export * from './useTableSelection';
-export * from './useTableSorting';
+export * from './useCustomCells';
+export * from './usePagination';
+export * from './useSelection';
+export * from './useSorting';
